+++ samples/AspNetCoreUsingDiagnostic/AspNetCoreUsingDiagnostic.csproj
@@ -12,7 +12,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AspNetCore.AutoHealthCheck\AspNetCore.AutoHealthCheck.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCore.AutoHealthCheck.Diagnostics\AspNetCore.AutoHealthCheck.Diagnostics.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCore.AutoHealthCheck\AspNetCore.AutoHealthCheck.csproj" />
   </ItemGroup>
 
 </Project>
