+++ samples/AspNetCoreUsingDiagnostic/Program.cs
