+++ samples/AspNetCoreUsingDiagnostic/appsettings.Development.json
