+++ samples/AspNetCoreUsingDiagnostic/appsettings.json
