+++ samples/WebApplication/Controllers/ValuesController.cs
@@ -57,14 +57,13 @@ namespace WebApplication.Controllers
         [HttpGet("complexGet/{id}")]
         public IActionResult ComplexGet(int id, [FromQuery] ComplexGetParam param)
         {
-            //throw new Exception();
             return Ok();
         }
 
         [HttpPost("array")]
         public IActionResult ComplexGet(List<ComplexGetParam> array)
         {
-            throw new Exception();
+            return new StatusCodeResult(500);
         }
     }
 }
