+++ samples/WebApplication/Plugins/ResultPlugin.cs
