+++ samples/WebApplication/Probes/CustomProbe.cs
