+++ samples/WebApplication/WebApplication.csproj
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AspNetCore.AutoHealthCheck\AspNetCore.AutoHealthCheck.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCore.AutoHealthCheck\AspNetCore.AutoHealthCheck.csproj" />
   </ItemGroup>
 
 </Project>
