+++ src/AspNetCore.AutoHealthCheck.Abstractions/Configurations/HealthCheckRuntimeMode.cs
@@ -20,18 +20,13 @@
 //SOFTWARE.
 // Project Lead - David Revoledo davidrevoledo@d-genix.com
 
-using System.Threading.Tasks;
-
 namespace AspNetCore.AutoHealthCheck
 {
-    internal interface IInternalRouteInformationEvaluator
+    /// <summary>
+    ///     Define how the internal runtime to auto execute the check will be called
+    /// </summary>
+    public enum HealthCheckRuntimeMode
     {
-        /// <summary>
-        ///     Internally evaluate a route to determine if needs to be included
-        ///     to check.
-        /// </summary>
-        /// <param name="routeInformation">route information</param>
-        /// <returns>if the route candidate needs to be included</returns>
-        Task<bool> Evaluate(IRouteInformation routeInformation);
+        Interval
     }
 }
