+++ src/AspNetCore.AutoHealthCheck.Abstractions/IAutoHealthCheckContext.cs
