+++ src/AspNetCore.AutoHealthCheck.Abstractions/IAutoHealthCheckContextAccessor.cs
