+++ src/AspNetCore.AutoHealthCheck.Abstractions/IHealthChecker.cs
