+++ src/AspNetCore.AutoHealthCheck.Abstractions/IProbesProcessor.cs
