+++ src/AspNetCore.AutoHealthCheck.Abstractions/IRouteDiscover.cs
