+++ src/AspNetCore.AutoHealthCheck.Abstractions/IRouteInformation.cs
@@ -1,0 +1,67 @@
+﻿//MIT License
+//Copyright(c) 2017 David Revoledo
+
+//Permission is hereby granted, free of charge, to any person obtaining a copy
+//of this software and associated documentation files (the "Software"), to deal
+//in the Software without restriction, including without limitation the rights
+//to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//copies of the Software, and to permit persons to whom the Software is
+//furnished to do so, subject to the following conditions:
+
+//The above copyright notice and this permission notice shall be included in all
+//copies or substantial portions of the Software.
+
+//THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//SOFTWARE.
+// Project Lead - David Revoledo davidrevoledo@d-genix.com
+
+using System;
+using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("AspNetCore.AutoHealthCheck")]
+
+namespace AspNetCore.AutoHealthCheck
+{
+    /// <summary>
+    ///     Route information for a single endpoint.
+    /// </summary>
+    public interface IRouteInformation
+    {
+        /// <summary>
+        ///     Http method needed to be consumed.
+        /// </summary>
+        string HttpMethod { get; }
+
+        /// <summary>
+        ///     Route needed to be consumed.
+        /// </summary>
+        string Path { get; }
+
+        /// <summary>
+        ///     Route template definition.
+        /// </summary>
+        string RouteTemplate { get; }
+
+        /// <summary>
+        ///     Route params key and type for url replacing.
+        /// </summary>
+        Dictionary<string, Type> RouteParams { get; }
+
+        /// <summary>
+        ///     Body params key and type
+        ///     Just 1 for now will be supported.
+        /// </summary>
+        Dictionary<string, Type> BodyParams { get; }
+
+        /// <summary>
+        ///     Query params key and type.
+        /// </summary>
+        Dictionary<string, Type> QueryParams { get; }
+    }
+}
