+++ src/AspNetCore.AutoHealthCheck.Abstractions/Plugins/IHealthCheckResultPlugin.cs
