+++ src/AspNetCore.AutoHealthCheck.Abstractions/Plugins/IHttpEndpointPlugin.cs
