+++ src/AspNetCore.AutoHealthCheck.Abstractions/Plugins/IPlugin.cs
