+++ src/AspNetCore.AutoHealthCheck.Abstractions/Probes/IProbe.cs
