+++ src/AspNetCore.AutoHealthCheck.Abstractions/Probes/ProbeResult.cs
