+++ src/AspNetCore.AutoHealthCheck.Abstractions/Results/HealthyResponse.cs
