+++ src/AspNetCore.AutoHealthCheck.Abstractions/Results/UnhealthyEndpoint.cs
