+++ src/AspNetCore.AutoHealthCheck.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sample", "sample", "{78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApplication", "WebApplication\WebApplication.csproj", "{B90A7A1C-6652-49B7-A75A-673003E054ED}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{DFA2E39C-2C56-42DE-8833-C36C392A07A7}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{270C7C3F-98AE-43D8-BB6B-18FAE0420F89}"
@@ -26,9 +24,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "solutionItems", "solutionIt
 		..\README.md = ..\README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetCoreUsingDiagnostic", "AspNetCoreUsingDiagnostic\AspNetCoreUsingDiagnostic.csproj", "{7BEDEB4A-9D89-4BFF-BE12-45C370A9A0D7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApplication", "..\samples\WebApplication\WebApplication.csproj", "{419C8FDE-17BA-42DF-AD2A-DF19323E706D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetCore.AutoHealthCheck.Diagnostics", "AspNetCore.AutoHealthCheck.Diagnostics\AspNetCore.AutoHealthCheck.Diagnostics.csproj", "{163F9A1A-813D-4E43-985A-3E8DD3D64B92}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreUsingDiagnostic", "..\samples\AspNetCoreUsingDiagnostic\AspNetCoreUsingDiagnostic.csproj", "{9E500F6B-4F0A-440E-8392-BB316C8D4048}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCore.AutoHealthCheck.Diagnostics", "AspNetCore.AutoHealthCheck.Diagnostics\AspNetCore.AutoHealthCheck.Diagnostics.csproj", "{A5D81A52-319E-489D-929E-5F285495938C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetCore.AutoHealthCheck.Abstractions", "AspNetCore.AutoHealthCheck.Abstractions\AspNetCore.AutoHealthCheck.Abstractions.csproj", "{7A4F34EB-A8AF-4A26-844F-DA2C52121622}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,32 +42,37 @@ Global
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B90A7A1C-6652-49B7-A75A-673003E054ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B90A7A1C-6652-49B7-A75A-673003E054ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B90A7A1C-6652-49B7-A75A-673003E054ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B90A7A1C-6652-49B7-A75A-673003E054ED}.Release|Any CPU.Build.0 = Release|Any CPU
 		{32641493-0310-4380-B7CF-47F55D74F0A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{32641493-0310-4380-B7CF-47F55D74F0A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{32641493-0310-4380-B7CF-47F55D74F0A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{32641493-0310-4380-B7CF-47F55D74F0A9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7BEDEB4A-9D89-4BFF-BE12-45C370A9A0D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7BEDEB4A-9D89-4BFF-BE12-45C370A9A0D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7BEDEB4A-9D89-4BFF-BE12-45C370A9A0D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7BEDEB4A-9D89-4BFF-BE12-45C370A9A0D7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{163F9A1A-813D-4E43-985A-3E8DD3D64B92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{163F9A1A-813D-4E43-985A-3E8DD3D64B92}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{163F9A1A-813D-4E43-985A-3E8DD3D64B92}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{163F9A1A-813D-4E43-985A-3E8DD3D64B92}.Release|Any CPU.Build.0 = Release|Any CPU
+		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{419C8FDE-17BA-42DF-AD2A-DF19323E706D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9E500F6B-4F0A-440E-8392-BB316C8D4048}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9E500F6B-4F0A-440E-8392-BB316C8D4048}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9E500F6B-4F0A-440E-8392-BB316C8D4048}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9E500F6B-4F0A-440E-8392-BB316C8D4048}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A5D81A52-319E-489D-929E-5F285495938C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A5D81A52-319E-489D-929E-5F285495938C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A5D81A52-319E-489D-929E-5F285495938C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A5D81A52-319E-489D-929E-5F285495938C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7A4F34EB-A8AF-4A26-844F-DA2C52121622}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7A4F34EB-A8AF-4A26-844F-DA2C52121622}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7A4F34EB-A8AF-4A26-844F-DA2C52121622}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7A4F34EB-A8AF-4A26-844F-DA2C52121622}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{B2553537-A52E-49A9-A36F-C42F5D441EEA} = {DFA2E39C-2C56-42DE-8833-C36C392A07A7}
-		{B90A7A1C-6652-49B7-A75A-673003E054ED} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
 		{32641493-0310-4380-B7CF-47F55D74F0A9} = {270C7C3F-98AE-43D8-BB6B-18FAE0420F89}
-		{7BEDEB4A-9D89-4BFF-BE12-45C370A9A0D7} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
-		{163F9A1A-813D-4E43-985A-3E8DD3D64B92} = {DFA2E39C-2C56-42DE-8833-C36C392A07A7}
+		{419C8FDE-17BA-42DF-AD2A-DF19323E706D} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
+		{9E500F6B-4F0A-440E-8392-BB316C8D4048} = {78D3E477-EE1C-4D6D-ADE9-366A4D7C6560}
+		{A5D81A52-319E-489D-929E-5F285495938C} = {DFA2E39C-2C56-42DE-8833-C36C392A07A7}
+		{7A4F34EB-A8AF-4A26-844F-DA2C52121622} = {DFA2E39C-2C56-42DE-8833-C36C392A07A7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {64B4BA9D-729D-4F8A-BBCF-F3E3FDA7E599}
